Refactor route handling for modularization
This update extracts API route definitions into separate files for better modularity and maintainability. Changes include consolidating `hono-openapi` route descriptions into dedicated `*.route.ts` files and revising corresponding logic in the main route files.
